move ever FormRequest inside Directory combined with them
