+++ app/Http/Controllers/Admin/TechnologyController.php
@@ -3,8 +3,8 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Models\Technology;
-use App\Http\Requests\StoreTechnologyRequest;
-use App\Http\Requests\UpdateTechnologyRequest;
+use App\Http\Requests\Admin\Technology\StoreTechnologyRequest;
+use App\Http\Requests\Admin\Technology\UpdateTechnologyRequest;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Str;
