+++ app/Http/Controllers/Admin/TypeController.php
@@ -3,8 +3,8 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Models\Type;
-use App\Http\Requests\StoreTypeRequest;
-use App\Http\Requests\UpdateTypeRequest;
+use App\Http\Requests\Admin\Type\StoreTypeRequest;
+use App\Http\Requests\Admin\Type\UpdateTypeRequest;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Str;
