+++ app/Http/Requests/Admin/Technology/StoreTechnologyRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Admin\Technology;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class StoreTypeRequest extends FormRequest
+class StoreTechnologyRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -22,8 +22,9 @@ class StoreTypeRequest extends FormRequest
     public function rules(): array
     {
         return [
-            'name' => 'required',
-            'description' => 'nullable',
+            'name' => 'required|unique:technologies,name',
+            'slug' => 'nullable',
+            'description' => 'nullable'
         ];
     }
 }
