+++ app/Http/Requests/Admin/Technology/UpdateTechnologyRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Admin\Technology;
 
 use Illuminate\Foundation\Http\FormRequest;
