+++ app/Http/Requests/Admin/Type/UpdateTypeRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Admin\Type;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class UpdatecollaboratorRequest extends FormRequest
+class UpdateTypeRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -23,8 +23,7 @@ class UpdatecollaboratorRequest extends FormRequest
     {
         return [
             'name' => 'required',
-            'url_git' => 'url|required',
-            'content' => 'nullable'
+            'description' => 'nullable',
         ];
     }
 }
